pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-agent"
version = "0.1.0"
description = "Gmail triage agent that uses Ollama and posts important messages to Discord"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests",
  "google-api-python-client",
  "google-auth",
  "google-auth-oauthlib",
]

[project.optional-dependencies]
dev = ["python-dotenv"]

[project.scripts]
email-agent = "email_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
